add context fixes
